Back out "RNTester-ios / RCTAppDelegate > correctly check for USE_HERMES Flag" (#41626)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/41626

In the codebase, we never set the RCT_USE_HERMES flag.

When we install the pods, we use the USE_HERMES flag and we set USE_HERMES as a build setting. So, the RCT_USE_HERMES flag will always be not set for the OSS.

https://pxl.cl/3RRxr

## Changelog:
[iOS][Fixed] - use the right USE_HERMES flag

## Facebook:
This change was incorrect as in OSS we never set the RCT_USE_HERMES flag, while we actually set the USE_HERMES one.

I will align the BUCK RNTester setup in the next diff of the stackog:

Reviewed By: dmytrorykun

Differential Revision: D51547810

fbshipit-source-id: 1da5b3a48a83a8ba49cf65382927bed2f9fc893d